+++ client/src/main/java/nl/tudelft/oopp/demo/entities/RestaurantHours.java
@@ -8,7 +8,7 @@ import java.util.Objects;
  */
 public class RestaurantHours {
     private int id;
-    private int day;
+    private long day;
     private Restaurant restaurant;
     private LocalTime startTime;
     private LocalTime endTime;
@@ -20,7 +20,7 @@ public class RestaurantHours {
      * @param startTime = the starting time
      * @param endTime = the ending time
      */
-    public RestaurantHours(int day, Restaurant restaurant, LocalTime startTime, LocalTime endTime) {
+    public RestaurantHours(long day, Restaurant restaurant, LocalTime startTime, LocalTime endTime) {
         this.day = day;
         this.restaurant = restaurant;
         this.startTime = startTime;
@@ -31,7 +31,7 @@ public class RestaurantHours {
 
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
 
@@ -51,7 +51,7 @@ public class RestaurantHours {
         return id;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
